startupPolicy: Emit event on disk source dropping
https://bugzilla.redhat.com/show_bug.cgi?id=769853

If a disk source gets dropped because it is not accessible,
mgmt application might want to be informed about this. Therefore
we need to emit an event. The event presented in this patch
is however a bit superset of what written above. The reason is simple:
an intention to be easily expanded, e.g. on 'user ejected disk
in guest' events. Therefore, callback gets source string and disk alias
(which should be unique among a domain) and reason (an integer);
(cherry-picked from commit baf2ff7e9054975becd7d87ef8c67eb817388e80)