Avoid generating the methods in multiple classes
The python code generator tries to figure out what class a
method should be in by looking at the list of arguments for
any which are object types. Unfortunately missing break
statements meant that methods which have multiple object
arguments (eg migrate as a virDomainPtr followed by a
virConnectPtr) got added to multiple classes.

The following incorrect methods are removed by this change

   virStream.download       (dup of virStorageVol.download)
   virStream.screenshot     (dup of virDomain.screenshot)
   virStream.upload         (dup of virStorageVol.upload)
   virConnect.migrate       (dup of virDomain.migrate)
   virConnect.migrate2      (dup of virDomain.migrate2)
   virConnect.migrate3      (dup of virDomain.migrate3)
   virConnect.migrateToURI3 (dup of virDomain.migrateToURI3)

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>