Skip copying manually written python for C APIs which don't exist
If the libvirt-override-virXXXX.py file has methods which call
C APIs that don't exist in the version of libvirt built against
we need to skip copying their code.

eg for 0.9.13 libvirt we should not copy the 'listAllDomains'
method.

The way this works is that it breaks the override file into
individual methods by looking for ' def '. It then collects
the contents until the next method start, whereupon it looks
for a libvirtmod.XXXXXX API call. It checks if the XXXXX part
is present in the XML description we have, and if not, it
discards the entire method.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>