sanitytest: Add virNetworkCreateXMLFlags() to list of name fixups
When checking whether each C API is exported to Python and vice
versa the sanitytest script is doing some name fixing. For
instance virNetworkCreateXML() is translated into
virConnect.networkCreateXML(). However, we have new C API on the
way: virNetworkCreateXMLFlags() which is not on the list for
these name fixups. Add it there.

Mind you, the python code generator works just fine because
generator.py:1082 compares just the prefix.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>