test_api_coverage: Reflect different naming in python for virDomainFDAssociate() and virDomainRestoreParams()
While virDomainRestoreParams() has 'virDomain' prefix (and thus
is put into Domain class), it is really in the same family as
virDomainRestore() or virDomainRestoreFlags() -> it acts upon
virConnect object and thus belongs into Connect class.

Then, virDomainFDAssociate is exposed as Domain.FDAssociate() but
because of the way we would generate the method's name
(fDAssociate) the test thinks it's not implemented.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>