spec: Set dist_name when creating a distribution tarball
When a distribution tarball is created (python3 -m build -n -s) the
dist_name in the spec file included in that tarball must match the
actual tarball name. The tarball name depends on the setuptools version
used to create it rather than on the version of setuptools installed on
the system RPMs are built.

Fixes: ed8ab5075d9190fdcd2894412d2127ac2a1d9535
Reported-by: Daniel P. Berrangé <berrange@redhat.com>
Signed-off-by: Jiri Denemark <jdenemar@redhat.com>